PrepAir-Backend-Main/requirements.txt
fastapi==0.95.0
uvicorn==0.23.0
motor==2.6.0
pydantic==1.10.2
bcrypt==4.0.1
PyJWT==2.8.0
random

question
options
correctOption
level
weightage
topic